requirements.txt
blinker==1.9.0
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.5
mariadb==1.1.11
MarkupSafe==3.0.2
mysql-connector-python==9.1.0
packaging==24.2
pycparser==2.22
python-dotenv==1.0.1
Werkzeug==3.1.3
